Refactor, we don't need container anymore
